requirements.txt
flask
pyserial
numpy
pillow
gpiozero
pixel-flipdot
flask-swagger
flask-swagger-ui
pydantic